pyproject.toml
[project]
name = "swipes-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "discord-py>=2.3.0",
    "python-dotenv>=1.2.2",
]
